requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
mysql-connector-python==9.0.0
python-dotenv==1.0.1
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
python-jose[cryptography]==3.3.0
email-validator==2.2.0
python-multipart==0.0.9